+++ tests/test_render_bond.py
@@ -18,7 +18,7 @@ def test_render_bond_cpu():
         }
     )
     system = System(data=data, box=Box([6.0, 6.0, 6.0], boundary=[0, 0, 0]))
-    system.cal_build_bond({("Cu", "Cu"): 0.5, ("Cu", "Zr"): 1.1, ("Zr", "Zr"): 1.1})
+    system.create_bonds({("Cu", "Cu"): 0.5, ("Cu", "Zr"): 1.1, ("Zr", "Zr"): 1.1})
 
     ren = TachyonRender(backend="cpu", antialiasing=False, ao=False)
     img = ren.render_system(